requirements.txt
python-telegram-bot
praw